Index variable `i` mistakenly shared within parallel region (#80)
The index `i` of the for-loop inside the parallel region (line 43) is shared by default. A `private(i)` clause may be necessary in the parallel construct to make sure that all threads are using their own index variable. Another possible fix is to write the for-loop at line 43 as `for (int i = 0; i < N; i++)`, which would allow to keep `i` shared in the parallel region and make sure the compiler is correctly making `i` firstprivate within the target-region. Let me know what you think. Thanks!