declare target directive requires at least one to or link clause
OpenMP 5.1 adds the following restriction:
> If the directive has a clause, it must contain at least one to clause or
> at least one link clause.

This restriction doesn't exist in OpenMP 5.0, but without a `to` clause
this test probably doesn't mean quite what it is supposed to:
> The form of the declare target directive that has no clauses and requires
> a matching end declare target directive defines an implicit extended-list
> to an implicit to clause.

Because the `declare target` directive in the test has a `device_type`
clause (and no `end declare target directive`), it is not of that form and
does not have an implicit `to`. Thus, the function `target_function` can
have a host version.