Fix mapping in target/test_target_defaultmap_to_from_tofrom.F90 (#541)
This testcase did spurious fail on some host/gpu arch combinations when
the Fortran pointer ptr wasn't nullified, either only for 'to' or
for 'from'. This commit does:
* Actually map the pointer (i.e. apply 'defaultmap(to/from/fromto) also to
  category 'pointer') instead of just using it inside the target.
* Initialize all variables to /= 0 to make checking work more properly.
* For 'tofrom', actually use the passed value to distinguish 'from' from
  'tofrom'.

Note: Regarding the pointer handling, see also the "restrictions to the
target construct" (last three bullet points in OpenMP 5.1/5.2). In
OpenMP 5.0, the third-last item describes this with looser words.