Issue 464: Avoid con't line issue in test_target_mapping_before_alloc.F90
When expanding macros with the C preprocessor, linebreaks are removed.
This causes for test_target_mapping_before_alloc.F90 a problem with a
'&' (linebreak) '&' continuation, which then ends up being mid line,
which is invalid Fortran.

As solution, the line break has been removed. This creates an overlong
but Fortran compilers have an option to accept this. (F202x will permit
longer than 132-characters free-source-form lines by default.)
This solution is in line with the similar issue, fixed
in Pull Req. 377 / commit 3586e45.

* tests/5.0/target/test_target_mapping_before_alloc.F90: Make the
  OMPVV_TEST_AND_SET_VERBOSE argument a single line and remove '&'
  continuation-line markers.