Issue 733: Fix test_requires_unified_address.c (#734)
* Issue 733: Fix test_requires_unified_address.c

Before this commit, the code assumed that the memory between host
and device is shared, which is not guaranteed. Change the code to
only check what the OpenMP spec guarantees for unified_address.

Namely, that is_device_ptr is not needed and that that host-side
pointer arithmetic is permitted for device pointers.

* test_requires_unified_address.c: Check function-return values

Diagnose failing allocation (and return early) and check also for
failing omp_target_memcpy.