Issue 733: More fixes for test_requires_unified_address.c (#737)
While using the device address inside the target region does not
require is_device_ptr, I missed for the previous code a special rule
for implicit mapping, namely for C/C++:

"A variable that is of type pointer [...] is treated as if it is the
 base pointer of a zero-length array section that had appeared as a
 list item in a map clause."

Thus, the previous test was invalid. Solution: Ensure the sizeof(void*)
bytes are actually available on the device and have the value of the host.
That can be either done via 'firstprivate' or by mapping the pointer.
This commit does either - once via 'defaultmap(firstprivate)' and once
via 'defaultmap(to)'.