Fixes for tests/5.0/target_update/*_discontiguous.c tests
This patch fixes two problems with the OpenMP 5.0 discontiguous 'target
update' tests: firstly, the length in an array section specifier does
not mean that "base+length" describes the exclusive upper limit for
the elements to be transferred when the stride is not one, but rather
it is the *number of elements* to be transferred.  (This is different
in Fortran, where the array syntax is (lower bound:upper bound:stride)
instead, and no elements beyond the upper bound will be accessed.)

That means that several tests contain buffer overruns, as they use
e.g. [0:N:2] for an array section specifier for an array of size N --
such an array section describes elements at indices 0, 2, 4, [...],
2*(N-1), which was clearly not the intention.

The patch also fixes an uninitialised 'error' count variable in
test_target_update_mapper_to_discontiguous.c.