test_metadirective_target_device_num.c: Add missing 'target' to #pragma omp
The first pragma lacked a 'target' in '#pragma omp enter data'; however,
adding one will enable the 'alloc' mapping of A such that the 'map(tofrom'
will not copy back data. Hence, an additional 'always' modifier is required.

(An alternative would be to remove the 'omp enter data' and the pointless
but valid 'omp target exit data'.)