test_target_memcpy_rect_async_{dep,no_}obj.F90: null ptr, fix idx calc
As pointer targets get automatically copied, 'fptr' either needs
to be disassociated (nullified) before mapping - or it could be
defined inside the target region (with 'block, ... end block').

Additionally there was an off-by-one issue with the pointer indices;
an alternative would be a two-dimensional array and using that
shape in 'c_f_pointer' call - and such leaving the calculation to
the compiler.