test_allocate_allocator.c: Fix 'omp allocate' usage
There were two issues with 'omp allocate':

First, the variable 'x' is declared ('int *x;') before the the
omp_allocator_handle_t variable 'x_alloc' is declared and initialized;
while the '#pragma omp allocate(x) allocator(x_alloc)' line comes
after the latter, there is a potential ordering issue:
Per C semantic, 'x' is available before 'x_alloc' but before it
can exist and be initialized, it has to be allocated - but for this
'x_alloc' has to exist.

Solution: Moving the 'x' declaration after the 'x_alloc' declaration
and initialization solves the issue.

Secondly, the OpenMP semantic states that the lifetime of an
omp-allocated automatic (stack) variable such as 'x' is its scope,
i.e. it ends at the enclosing '}'. However, as allocator used for
allocation is also used for deallocation, its lifetime must exceed the
liftime of 'x'. Thus, the 'omp_destory_allocator' cannot be in in the
same scope as 'x'.

Solution: Place the omp-allocated variables in an inner scope such
that 'x_alloc' has a longer lifetime.

Additionally, some comments have been added to make it clearer to
the code reader what happens and why code is as it is.