Fix test_target_defaultmap_present_pointer.F90
Follow up to Pull Request #711, which added this testcase.

The problem is that using

    !$omp target enter data map(to: ptr)
    !$omp target map(tofrom: errors) defaultmap(present:pointer)
        ...
        ptr(i) = i + 2
        ...
    !$omp end target
    !$omp target exit data map(delete: ptr)

the data is not copied back to the device. This will only work if either
host fallback is done - or 'map(to: ptr)' is a _self_ map.
Solution is to copy it back, e.g. 'map(from:' as done in this commit
or '!$omp target update from(ptr)'.