Issue 855: Fix make.def by not force-override user compiler settings (#856)
This fixes a regression caused by commit f89b8a4 (Jan 9, 2025).

The idea of that commit seems to permit for updating the settings using
the default settings - and in order to do so, the compiler name without
path ("nodir") is required for the pattern matching.

However, the current implementation overrides the name completely such
instead of '/path/to/compiler/cc' the compiler 'cc' is used. As the
commit uses "override" there is no way for a user - neither on the
command line nor by setting the value earlier - to prevent this.

As the purpose seems to be only to match the pattern, this commit
introduces a new variable (whose name shouldn't clash) that is only
used for the compiler-name pattern match. That way, the specified
full path of the compiler is used.

However, if not only the compiler path but also all setting should not be
overridden, more is needed. This commit introduces in addition the make
variable OMPVV_NO_COMPILER_CHANGES that, if set, prevents the
code changes of make.def - while still adding the header file directory
and the escape sequences.