Issue 863: test_interop_target.c - remove nowait from 'interop destroy' to fix dependency (#867)
The 'depend' clause clause is required to avoid the race between writing the data
(before the 'interop' directive) and reading it without further synchronisation
after the interop destroy line.

However, with 'nowait' it handles the tracking in the background to handle the
dependency correctly for the streaming object (targetsync) but as it is
run asynchronously, the host code proceeds to the reading code - and failed.

By removing the 'nowait', the host code is only excuted after the stream
(with dependency tracking) is processed; the actual stream processing should
take no time as no actual stream processing is happening.