Issue 896: Fix race in 5.0/teams_loop/test_target_teams_loop_allocate.c
The code uses 'target teams loop' with 'firstprivate(local)'; however, as
'firstprivate' does not apply to 'loop', it is only privatized at the
contention group level and all loop threads share the same 'local' variable,
causing races when 'loop' is executed thread parallel.

The solution is to use a workdistribution loop (aka 'for'), which also
accepts a 'firstprivate'; as 'teams for' is not permitted, that's
'target teams distribute parallel for'.

Technically, a 'private' would have done as well as the original value is
never read, but this patch keeps the 'firstprivate'. Note that while 'loop'
accepts neither 'allocate' nor 'firstprivate' (albeit 'private' is accepted),
all those clauses are accepted by the workdistribution-loop 'for' construct.