Avoid race in 5.0/teams_loop/test_target_teams_loop_allocate.c differently
Review comments suggested that having the allocator only applicable to 'teams'
and combining with 'loop' is crucial. This new attempt uses 'thread_limit(1)'
to ensure that there is only a single thread per team (contention group),
avoiding a race due to the 'loop' but still permitting to use a scalar that
is only privatized at teams level.